feat: refactored codes

+++ FastFoodWebApi/DTOs/OrderCreateDto.cs
@@ -32,7 +32,7 @@ namespace FastFoodWebApi.DTOs
         public int CustomerId { get; set; }
 
 
-        public float DeliveryCharge { get; set; }
+        public decimal DeliveryCharge { get; set; }
         public TimeSpan EstimateTime { get; set; }
 
         [MaxLength(50)]
